+++ AtlasLoot_WorldEvents/data.lua
@@ -1396,7 +1396,13 @@ data["HarvestFestival"] = {
 				{ 2,  19996 },                                                                                                                              -- Harvest Fish
 				{ 3,  19994 },                                                                                                                              -- Harvest Fruit
 				{ 4,  19997 },                                                                                                                              -- Harvest Nectar
-				{ 16, [ATLASLOOT_IT_ALLIANCE] = 20009, [ATLASLOOT_IT_HORDE] = 20010, [QUEST_EXTRA_ITTYPE] = (UnitFactionGroup("player") == "Horde") and 8150 or 8149 }, -- For the Light! / The Horde's Hellscream
+				{ 6,  34850,                           [PRICE_EXTRA_ITTYPE] = "money:85" },                                                                 -- Midsummer Ground Flower
+				{ 7,  9312,                            [PRICE_EXTRA_ITTYPE] = "money:17" },                                                                 -- Blue Firework
+				{ 8,  9313,                            [PRICE_EXTRA_ITTYPE] = "money:17" },                                                                 -- Green Firework
+				{ 9,  9318,                            [PRICE_EXTRA_ITTYPE] = "money:17" },                                                                 -- Red Firework
+				{ 10, 9314,                            [PRICE_EXTRA_ITTYPE] = "money:43" },                                                                 -- Red Streaks Firework
+				{ 11, 9315,                            [PRICE_EXTRA_ITTYPE] = "money:34" },                                                                 -- Yellow Rose Firework
+				{ 16, [ATLASLOOT_IT_ALLIANCE] = 20009, [ATLASLOOT_IT_HORDE] = 20010,     [QUEST_EXTRA_ITTYPE] = (UnitFactionGroup("player") == "Horde") and 8150 or 8149 }, -- For the Light! / The Horde's Hellscream
 				{ 17, 19697 },                                                                                                                              -- Bounty of the Harvest
 				{ 18, 19696 },                                                                                                                              -- Harvest Bread
 			},